pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kolett"
version = "0.1.0"
description = "An open-source, database-agnostic delivery tool for VFX packages."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Hristo Velev", email = "hristo@bottleshipvfx.com" }
]
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
kolett = "kolett.runner:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
"Homepage" = "https://github.com/HristoVelev/kolett"
"Bug Tracker" = "https://github.com/HristoVelev/kolett/issues"
